requirements.txt
aiomysql==0.3.2
annotated-types==0.7.0
anyio==3.7.1
certifi==2025.10.5
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
fastapi==0.104.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.25.1
idna==3.11
Pillow==10.1.0
pycparser==2.23
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
PyMySQL==1.1.0
python-dotenv==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlmodel==0.0.27
starlette==0.27.0
typing_extensions==4.15.0
uvicorn==0.24.0
